chart implementation with chart.js and plotly

+++ expenses/income/models.py
@@ -6,7 +6,7 @@ from datetime import datetime
 # Create your models here.
 class Income(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    amount = models.DecimalField(max_digits=9, decimal_places=2)
+    amount = models.FloatField()
     source = models.CharField(max_length=255)
     description = models.TextField(blank=True,null=True)
     date = models.DateTimeField(default= datetime.now)
